Update SWR when searching.
